% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{app_theme}
\alias{app_theme}
\title{Defines the dashboard theme for the Vanishing Vitamin Shiny app}
\usage{
app_theme()
}
\value{
the result of \code{fresh::create_theme}
}
\description{
This is an internal function that is used within the exported launch_app()
function.
}
