% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{citations}
\alias{citations}
\title{Citation information for published data}
\format{
\subsection{\code{citations}}{

A tibble object with 15 rows and 4 columns
\describe{
\item{unique_id}{chr, string representing the source of the data. If the data are published, this wil be the DOI of the associated article. If the data are not published, this will be a string representing the collection location.}
\item{metadata}{A nested column. Each element contains metadata about the associated publication. Has value "NA" for unpublished data.}
\item{formatted_metadata}{A formatted character string containing information stored in the metadata column. Used for display in the app.}
}
}
}
\usage{
citations
}
\description{
Citation information for published data
}
\keyword{datasets}
