% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{dose_response}
\alias{dose_response}
\title{Non-exported helper function for computing dose response}
\usage{
dose_response(Thiamin_conc, ec50_mu, slope_p, upper_p, lower_p = 0)
}
\description{
Non-exported helper function for computing dose response
}
