% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lc50_curve}
\alias{lc50_curve}
\title{Model-based LC50 data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6548 rows and 14 columns.
}
\usage{
lc50_curve
}
\description{
Model-based LC50 data
}
\keyword{datasets}
