% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Defines the Server function for the Vanishing Vitamin Shiny app}
\usage{
app_server()
}
\value{
a function object containing app server logic
}
\description{
This is an internal function that is used within the exported launch_app()
function.
}
