% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Defines the User Interface elements for the Vanishing Vitamin Shiny app}
\usage{
app_ui(tdc_data)
}
\arguments{
\item{tdc_data}{data set containing Thiamin by Survivability data. Should be
the data set exported by the vanishingVitamin package,
vanishingVitamin::tdc_data}
}
\value{
\code{bs4Dash::dashboardPage} object containing UI elements
}
\description{
This is an internal function that is used within the exported launch_app()
function.
}
