% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{citations}
\alias{citations}
\title{Citation information for published data}
\format{
\subsection{\code{citations}}{

A tibble object with 15 rows and 4 columns
\describe{
\item{DOI}{URL associated with the published data. May be a URL or "unpublished" for unpublished data.}
\item{DOI_num}{DOI associated with the published data. May be a URL or "unpublished" for unpublished data.}
\item{metadata}{A nested column. Each element contains metadata about the associated publication. Has value "NA" for unpublished data.}
\item{formatted_metadata}{A formatted character string containing information stored in the metadata column. Used for display in the app.}
}
}
}
\usage{
citations
}
\description{
Citation information for published data
}
\keyword{datasets}
