% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dose_response_params}
\alias{dose_response_params}
\title{Dose response curve fitted parameter values.}
\format{
\subsection{\code{dose_response_params}}{

A tibble object with 108 rows and 5 columns:

\describe{
\item{species}{chr, species associated with parameter value. One of: Atlantic, Chinook, Coho, Lake Trout, Steelhead}
\item{region}{chr, region associated with parameter value. One of: Great Lakes, Pacific, Baltic}
\item{parameter}{chr, name of parameter in dose response model. One of: ec50_mu, ec50_sigma, slope_p, upper_p}
\item{median}{num, estimated posterior median (50th percentile) value of associated parameter}
\item{CI_2.5}{num, estimated posterior 2.5th percentile value of associated parameter}
\item{CI_97.5}{num, estimated posterior 97.5th percentile value of associated parameter}
}
}
}
\usage{
dose_response_params
}
\description{
Fitted values generated by Miles Daniels (miles.daniels@usgs.gov).

Species-specific parameters are available for Atlantic, Chinook, Coho, Lake Trout, and Steelhead

Each species has an associated region, indicating where data were collected.
Coho, Lake Trout, and Steelhead only have one region: Great Lakes.
Atlantic and Chinook have both overall and region-specific parameters.

Atlantic: Baltic and Great Lakes

Chinook: Pacific and Great Lakes

Each row of the data set represents a fitted parameter values for a combination of
species, region, model parameter, and parameter quantile.
Parameter quantiles are used to visualize the median dose response curves and 95\% credible interval.
}
\keyword{datasets}
