% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Vanishing Vitamin app}
\usage{
launch_app(options = list())
}
\arguments{
\item{options}{list; app launch options passed to \code{options} argument of
shiny::shinyApp()}
}
\description{
Launch Vanishing Vitamin app
}
\examples{
launch_app()


}
\seealso{
\code{\link[shiny:shinyApp]{shiny::shinyApp()}}
}
