% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lc50_curve}
\alias{lc50_curve}
\title{Model-based LC50 data}
\format{
\subsection{\code{lc50_curve} A tibble object with 6,548 rows and 14 columns}{
}
}
\usage{
lc50_curve
}
\description{
These data were generated based on a sigmoidal Bayesian dose-response model
developed by Miles Daniels that models the relationship between Thiamin
concentration and survival rate among fish. The model was fit to the
\code{tdc_data} data set also exported by the VanishingVitamin package, resulting
in estimates for the 3 model parameters and one response variable (survival
rate).
}
\details{
This data set summarizes quantiles of these model estimates for the sake of
constructing point and interval estimates of survival rate given thiamin
concentration. See the "See also" section for a link to more details about the
model. An unexported function, \code{vanishingVitamin:::dose_response()}, was
used to compute the estimated survival rate given the parameter values and
thiamin concentration.

These data are used to create a median dose response curve with approximate
95\% credible intervals. The \code{Thiamin_conc} contain abcissa ('x') values
and the three \code{survival_*} columns contain ordinate ('y') values for the
0.25, 0.5 (i.e., median), and 0.975 quantiles of estimated survival,
respectively.
}
\seealso{
\url{https://github.com/MilesEDaniels/Thiamine-Dependent-Fry-Mortality?tab=readme-ov-file#methods}

\describe{
\item{ec50_2.5}{ the 0.025 quantile of the EC50 parameter}
\item{ec50_50}{ the 0.5 quantile (i.e., median) of the EC50 parameter}
\item{ec50_97.5}{ the 0.975 quantile of the EC50 parameter}
\item{slope_2.5}{ the 0.025 quantile of the slope (S) parameter}
\item{slope_50}{ the 0.5 quantile (i.e., median) of the slope (S) parameter}
\item{slope_97.5}{ the 0.975 quantile of the slope (S) parameter}
\item{upper_2.5}{ the 0.025 quantile of the upper (U) parameter}
\item{upper_50}{ the 0.5 quantile (i.e., median) of the upper (U) parameter}
\item{upper_97.5}{ the 0.975 quantile of the upper (U) parameter}
\item{Thiamin_conc}{ thiamin concentration values. This is the independent variable in the model.}
\item{survival_ci_2.5}{ 0.025 quantile of the estimated survival, given the 0.025 parameter quantiles and a thiamin concentration}
\item{survival_median}{ 0.5 quantile (i.e., median) of the estimated survival, given the 0.5 parameter quantiles and thiamin concentration}
\item{survival_ci_97.5}{ 0.975 quantile of the estimated survival, given the 0.975 parameter quantiles and thiamin concentration}
\item{plot_label}{ a character string of formatted text summarizing the estimated survival rate values for a given thiamin concentration. Meant to be used for a tooltip in an interactive scatterplot.}
}
}
\keyword{datasets}
